Revert data source changes
This reverts a commit that was made before the project was squashed and
open sourced removing the concept of data sources due to changing
requirements in an effort to simply the code. A significant amount of
the work to support operation archive has already been completed and we
can keep it around in a branch, with a rever of this commit, as the base
in case we decide to pick up the feature again.