make async markdown renderer (now) synchronous (#59)
This is a followup to #41 and may actually allow us to re-enable syntax highlighting. #41 didn't address the real issue which was that the markdown renderer was loading async after evidence was queued up. This caused the requests to block and the rendered to not get called until after all evidence was loaded, which could take a long time. This eagerly loads the rendered before files are requested so that the descriptions can be rendered while the files are being requested.