Change dependabot to weekly
Set the dependabot interval for npm and go to weekly due to the
frequency of changes for specific many of the packages. These typically
are small point releases many of which are unimportant. This seems to be
handled differently for security related issues. We may choose to push
this out to monthly even depending on how loud this still is.