Round times to the nearest second and chop nanos
Clockwork changed to defaulting to the current time rather than a fixed
time. This means that the nanoseconds field is now set which isn't the
case in the MySQL timestamp. The value returned from MySQL is always
going to be limited to the nearest second but the value in Go that is
expected will have nanosecond precision. We need to chop it off.