Update packages
This updates a series of packages related to eslint, react, and
typescript. Major version bumps required a whole bunch of packages to
all be updated which, even after months, we're still waiting for some.
The last one needed is eslint-plugin-react-hooks but we can use a
pre-release of it for now until version 5.1 is ready