Add OIDC client specific urls
This optionally allows you to specify the the url of the ashirt server
to redirect to after login specific to the OIDC client. This is
necessary in environments where you may have two different domains
attached to the same instance with two different OIDC configs (eg. a
corp one and an OOB one). This is sort of a hacky solution and we should
be able to create a more automated one but we'll likely need to redesign
some of the auth stack first to do so.